Demo Security WebTestClient support
We can run our tests with an actual user by authenticating the request.
However, this couples our logic to which users currently exist. This
is not ideal when we are just wanting to test logic other than
authentication (i.e. authorization).

If we are not concerned with authentication for our test we can use
withUser() to run as a specific user when running as WebTestClient in
a mocked request (i.e. it won't work for full integration tests)